feat: check phase before backend operations

+++ server/src/dao/proposal.dao.mjs
@@ -1,6 +1,7 @@
 "use strict";
 
 import db from "../db/db.mjs";
+import GP from "./utils.dao.mjs";
 
 class ProposalDAO {
   /**
@@ -54,25 +55,30 @@ class ProposalDAO {
    * @returns
    */
   async insertProposal(user, proposal) {
-    return new Promise((resolve, reject) => {
-      // check if the user has already submitted 3 proposals
-      const sqlCheck = `SELECT * FROM proposals WHERE author = ?`;
-      db.all(sqlCheck, [user.username], (err, rows) => {
-        if (err) reject(err);
-        // user can submit a proposal
-        else if (!rows || rows.length < 3) {
-          const sql = `INSERT INTO proposals (author, description, cost)VALUES (?, ?, ?)`;
-          db.run(
-            sql,
-            [user.username, proposal.description, proposal.cost || 0],
-            (err) => {
-              if (err) reject(err);
-              else resolve(true);
-            }
-          );
-          // user has already submitted 3 proposals
-        } else reject(new Error("You have already submitted 3 proposals"));
-      });
+    return new Promise(async (resolve, reject) => {
+      // check if it is the first phase
+      const phase = await GP();
+      if (phase.phase !== 1) reject(new Error("Not in phase 1"));
+      else {
+        // check if the user has already submitted 3 proposals
+        const sqlCheck = `SELECT * FROM proposals WHERE author = ?`;
+        db.all(sqlCheck, [user.username], (err, rows) => {
+          if (err) reject(err);
+          // user can submit a proposal
+          else if (!rows || rows.length < 3) {
+            const sql = `INSERT INTO proposals (author, description, cost)VALUES (?, ?, ?)`;
+            db.run(
+              sql,
+              [user.username, proposal.description, proposal.cost || 0],
+              (err) => {
+                if (err) reject(err);
+                else resolve(true);
+              }
+            );
+            // user has already submitted 3 proposals
+          } else reject(new Error("You have already submitted 3 proposals"));
+        });
+      }
     });
   }
 
@@ -82,26 +88,31 @@ class ProposalDAO {
    * @param proposal - The proposal to be edited.
    */
   async editProposal(user, proposal) {
-    return new Promise((resolve, reject) => {
-      // check if the user is the author of the proposal
-      const sqlCheck = `SELECT * FROM proposals WHERE id = ?`;
-      db.get(sqlCheck, [proposal.id], (err, row) => {
-        if (err) reject(err);
-        else if (!row) reject(new Error("Proposal not found"));
-        else if (user.username !== row.author)
-          reject(new Error("User cannot edit another user's proposal"));
-        else {
-          const sql = `UPDATE proposals SET description = ?, cost = ? WHERE id = ?`;
-          db.run(
-            sql,
-            [proposal.description, proposal.cost, proposal.id],
-            (err) => {
-              if (err) reject(err);
-              else resolve(true);
-            }
-          );
-        }
-      });
+    return new Promise(async (resolve, reject) => {
+      // check if it is the first phase
+      const phase = await GP();
+      if (phase.phase !== 1) reject(new Error("Not in phase 1"));
+      else {
+        // check if the user is the author of the proposal
+        const sqlCheck = `SELECT * FROM proposals WHERE id = ?`;
+        db.get(sqlCheck, [proposal.id], (err, row) => {
+          if (err) reject(err);
+          else if (!row) reject(new Error("Proposal not found"));
+          else if (user.username !== row.author)
+            reject(new Error("User cannot edit another user's proposal"));
+          else {
+            const sql = `UPDATE proposals SET description = ?, cost = ? WHERE id = ?`;
+            db.run(
+              sql,
+              [proposal.description, proposal.cost, proposal.id],
+              (err) => {
+                if (err) reject(err);
+                else resolve(true);
+              }
+            );
+          }
+        });
+      }
     });
   }
 
@@ -111,12 +122,17 @@ class ProposalDAO {
    * @returns true if no error occurs
    */
   async removeProposal(proposal) {
-    return new Promise((resolve, reject) => {
-      const sql = `DELETE FROM proposals WHERE id = ?`;
-      db.run(sql, [proposal.id], (err) => {
-        if (err) reject(err);
-        else resolve(true);
-      });
+    return new Promise(async (resolve, reject) => {
+      // check if it is the first phase
+      const phase = await GP();
+      if (phase.phase !== 1) reject(new Error("Not in phase 1"));
+      else {
+        const sql = `DELETE FROM proposals WHERE id = ?`;
+        db.run(sql, [proposal.id], (err) => {
+          if (err) reject(err);
+          else resolve(true);
+        });
+      }
     });
   }
